feat: asking only once in twenty mins of reopening the project.
If we open a same project multiple of times in small time period,
we get asked the same question to pull or fetch each time. For that
we now managed to shut it for another 20 mins to not ask the
question once it asked. This 20 mins are individual for each
project or git repo.